==> CODES/01_job-post__treating-missing-values.Rmd <==
---
title: "[Thesis Draft] Treating missing values"
author: "Thien Binh Hoang Dao"
date: "`r format(Sys.time(), '%Y-%m-%d %H:%M')`"
output:
  html_document: 
    keep_md: false
---

```{r setup, include = F}

knitr::opts_chunk$set(echo = F, out.width = "90%", fig.width = 12, fig.height = 8, warning = F, message = F)

```

``` {r}

###############################################
## Code: 01_job-post__treating-missing-values
## Purpose: Treating missing data
## Author: Thien Dao
## Creation date: 2021-04-09
## Update date: 2021-04-22
###############################################

```

``` {r load-lib}

########################
# Loading libraries
########################

library(dplyr)
library(stringr)
library(DT)
library(naniar)
library(ggplot2)
library(Hmisc)
library(purrr)

```

``` {r import-data}

########################
# Importing raw data
########################

base_path <- "E:\\THIENDHB_GOOGLEDRIVE\\MASTER TILBURG\\THESIS/"
df <- read.csv(file.path(base_path, "DATASET/INPUT/online-job-postings.csv"), stringsAsFactors = F)
df <- df %>% mutate(job_id = 1:n())
refined_df <- df

```

``` {r functions}
########################
# Functions
########################
# Helper function to detect missing values and short words
detect_na <- function(x) {
  is.na(x) |
    x == "na" |
    x == "NA" |
    x == "NaN" |
    x == "n/a" | 
    x == "N/A" |
    x == ""
}

detect_na_full <- function(x) {
  is.na(x) |
    x == "na" |
    x == "NA" |
    x == "NaN" |
    x == "n/a" | 
    x == "N/A" |
    x == "N /A" |
    x == "N / A" |
    x == "N/ A" |
    x == "N/A." |
    x == "H/A" |
    x == "N/" |
    x == ""
}
 
# # Helper function to detect short words
# # (short word means all member words have length = 1 after splitting by punctuation)
# detect_shortwords <- function(x) {
#     sum(str_length(str_split(
#       gsub('[[:punct:]]', ' ', x), ' ', simplify = T
#     )) > 1) == 0                 
# }

```

``` {r }

summary(df[c('jobpost', 'Title', 'JobDescription', 'JobRequirment', 'RequiredQual')])

```

``` {r describe-data}

df2 <- df[c('jobpost', 'Title', 'JobDescription', 'JobRequirment', 'RequiredQual')]
df2[detect_na_full(df2)] <- NA

classes <- sapply(df2, class, simplify = T)

describe_tbl <- map_df(describe(df2), 
                       ~.x$counts, 
                       .id = "Column")

describe_tbl <- describe_tbl %>% 
  select(Column, Observations = n, "Unique values" = distinct) %>%
  mutate("Data type" = classes) %>%
  select(Column, "Data type", everything())

datatable(describe_tbl)

describe_latex <- Hmisc::latex(describe_tbl,
                               caption = 'Ininitial dataset description',
                               label = 'tbl:data_description',
                               rowname = NULL,
                               file = paste(base_path, '/RESULTS/dataset_description', ".tex", sep="")
)

```

``` {r explore-data}

# datatable(df[c('jobpost', 'Title', 'JobDescription', 'JobRequirment', 'RequiredQual')], escape=T)

# Detect and convert all considered "missing" values to NA
refined_df[detect_na(refined_df)] <- NA

```

``` {r visualize-b4}

########################
# Visualizing missing data
########################

# Visualize missing data
vis_miss(refined_df[c('Title', 'JobDescription', 'JobRequirment', 'RequiredQual')]) +
  labs(title = "Row-wise missingness of Online Job Posts data")

try({
  missplot <- gg_miss_upset(
  refined_df[c('Title', 'JobDescription', 'JobRequirment', 'RequiredQual')],
  matrix.color = "#ff9800",
  main.bar.color = "#900c3f",
  sets.bar.color = "#009933",
  set_size.show = T,
  set_size.scale_max = 4500
)
  print(missplot)
  
  ggsave(
  filename = paste0(base_path, "RESULTS/initial_missingness.png"),
  plot = print(missplot)
)
  },
silent = T)

```

``` {r treat-missing-data}

########################
# Treating missing data
# Method:
# - Find corresponding section in column jobpost
# - Extract that section into new section column
# - Only treat sections of Job title, description, requirement, and qualification required
########################

# Dealing with missing job title
#==============================================

titles <- refined_df %>% filter(is.na(Title))

start_str <- str_locate(titles$jobpost, '(?<=TITLE:)')[,1]
end_str <- str_locate(str_sub(titles$jobpost, start=start_str), '\n[A-Z/\\ ]{1,}:')[,1] + start_str - 2
end_str <- ifelse(!is.na(start_str) & is.na(end_str), str_length(job_desc$jobpost), end_str)

extract_title <- str_trim(str_sub(titles$jobpost, start_str, end_str))
titles <- titles %>%
  mutate(new_title = extract_title)
# datatable(titles[c('jobpost', 'Title', 'new_title')], escape=T)

refined_df <- left_join(refined_df, select(titles, job_id, new_title), by = "job_id")
refined_df <- refined_df %>%
  mutate(Title = case_when(!is.na(new_title) ~ new_title, TRUE ~ Title))

# Dealing with missing job description
#==============================================

job_desc <- refined_df %>% filter(is.na(JobDescription))

start_str <- str_locate(job_desc$jobpost, '(?<=DESCRIPTION:|NEWS DETAILS:)')[,1]
end_str <- str_locate(str_sub(job_desc$jobpost, start=start_str), '\n[A-Z/\\ ]{1,}:')[,1] + start_str - 2
end_str <- ifelse(!is.na(start_str) & is.na(end_str), str_length(job_desc$jobpost), end_str)

extract_desc <- str_trim(str_sub(job_desc$jobpost, start_str, end_str))
job_desc <- job_desc %>%
  mutate(new_desc = extract_desc, start_desc = start_str, end_desc = end_str)
# datatable(job_desc[c('jobpost', 'JobDescription', 'new_desc', 'start_desc', 'end_desc

refined_df <- left_join(refined_df, select(job_desc, job_id, new_desc), by = "job_id")
refined_df <- refined_df %>%
  mutate(JobDescription = case_when(!is.na(new_desc) ~ new_desc, TRUE ~ JobDescription))

# Dealing with missing job requirement
#==============================================

job_req <- refined_df %>% filter(is.na(JobRequirment))

start_str <- str_locate(job_req$jobpost, '(?<=RESPONSIBILITIES:)')[,1]
end_str <- str_locate(str_sub(job_req$jobpost, start=start_str), '\n[A-Z/\\ ]{1,}:')[,1] + start_str - 2
end_str <- ifelse(!is.na(start_str) & is.na(end_str), str_length(job_req$jobpost), end_str)

extract_req <- str_trim(str_sub(job_req$jobpost, start_str, end_str))
job_req <- job_req %>%
  mutate(new_req = extract_req)
refined_df <- left_join(refined_df, select(job_req, job_id, new_req), by = "job_id")
refined_df <- refined_df %>%
  mutate(JobRequirment = case_when(!is.na(new_req) ~ new_req, TRUE ~ JobRequirment))

# Dealing with missing required qualification
#==============================================

job_qual <- refined_df %>% filter(is.na(RequiredQual))

start_str <- str_locate(job_qual$jobpost, '(?<=QUALIFICATIONS:)')[,1]
end_str <- str_locate(str_sub(job_qual$jobpost, start=start_str), '\n[A-Z/\\ ]{1,}:')[,1] + start_str - 2
end_str <- ifelse(!is.na(start_str) & is.na(end_str), str_length(job_qual$jobpost), end_str)

extract_qual <- str_trim(str_sub(job_qual$jobpost, start_str, end_str))
job_qual <- job_qual %>%
  mutate(new_qual = extract_qual)
refined_df <- left_join(refined_df, select(job_qual, job_id, new_qual), by = "job_id")
refined_df <- refined_df %>%
  mutate(RequiredQual = case_when(!is.na(new_qual) ~ new_qual, TRUE ~ RequiredQual))

```

``` {r visualize-after-retrieval}

########################
# Visualizing after-retrieval missing data
########################

# Detect and convert all considered "missing" values to NA
refined_df[detect_na_full(refined_df)] <- NA

vis_miss(refined_df[c('Title', 'JobDescription', 'JobRequirment', 'RequiredQual')]) +
  labs(title = "Row-wise missingness of Retrieval Online Job Posts data")

try({
  missplot <- gg_miss_upset(
  refined_df[c('Title', 'JobDescription', 'JobRequirment', 'RequiredQual')],
  matrix.color = "#ff9800",
  main.bar.color = "#900c3f",
  sets.bar.color = "#009933",
  set_size.show = T,
  set_size.scale_max = 4500
)
  print(missplot)
  
  ggsave(
  filename = paste0(base_path, "RESULTS/retrieval_missingness.png"),
  plot = print(missplot)
)
  },
silent = T)

```

``` {r replace-missing}

########################
# Remove useless rows
# (missing all description, requirement, and qualification)
########################

new_df <- refined_df %>%
  filter(!(is.na(JobDescription) & is.na(JobRequirment) & is.na(RequiredQual)))

########################
# Replace missing values "NA" by "Unprovided"
########################

new_df[is.na(new_df)] <- "Unprovided"

```

``` {r filter-columns}

########################
# Keep and rename potential useful columns
########################

new_df<- new_df %>%
  select(job_id,
         job_title = Title,
         job_description = JobDescription,
         job_requirement = JobRequirment,
         job_qualification = RequiredQual)

```

``` {r}

describe_df <- map_df(describe(new_df %>% mutate(job_id = as.character(job_id))), 
                      ~.x$counts, 
                      .id = "Column")

imputed <- sapply(new_df, function(x) {sum(x == 'Unprovided')}, simplify = T)

old_columns <- c('row number', 'Title', 'JobDescription', 'JobRequirment', 'RequiredQual')

describe_final_df <- describe_df %>% select(Column, Observations = n, "Unique values" = distinct) %>%
  mutate("Imputed values" = imputed,
         "Old column name" = old_columns) %>%
  select(Column, "Old column name", everything())

datatable(describe_final_df)

latex_df <- Hmisc::latex(describe_final_df,
                               caption = 'Final dataset description',
                               label = 'tbl:final_data_description',
                               rowname = NULL,
                               file = paste(base_path, 'RESULTS/final_data_description', ".tex", sep="")
)

```

``` {r re-investigate-data}

# datatable(new_df, escape=T)

```

``` {r visualize-final}

########################
# Visualizing after-replacement missing data
########################

vis_miss(new_df[c('job_title', 'job_description', 'job_requirement', 'job_qualification')]) +
  labs(title = "Row-wise missingness of Refined Online Job Posts data")

try(
  gg_miss_upset(new_df[c('job_title', 'job_description', 'job_requirement', 'job_qualification')],
                matrix.color = "#ff9800", main.bar.color = "#900c3f", sets.bar.color = "#009933",
                empty.intersections = T, set_size.show = T),
  silent = T
)

```

``` {r export}

########################
# Export final output to CSV
########################

write.csv(new_df, file.path(base_path, "DATASET/INPUT/refined_jobpost_data.csv"), row.names = F)

```
