==> CODES/03_job-skills_02c_doc_embeddings.ipynb <==
## Load libraries

# Libraries to work cross-platform
import os

# Libraries to work with dataset
import numpy as np
import pandas as pd
import ast                      # convert string to list after importing csv data
# import dask.dataframe as dd
import pickle

# Libraries to pre-process data
# import spacy
from gensim.models.keyedvectors import KeyedVectors

# Libraries to visualize data
# from wordcloud import WordCloud
# import matplotlib.pyplot as plt
# import seaborn as sns

# Libraries for monitoring operation process
from tqdm import tqdm

## Configurate and declare global variables

os_name = os.name

if os_name == 'nt':
    BASE_DIR = "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/"
    INPUT_DIR = BASE_DIR + "DATASET/INPUT/"
    OUTPUT_DIR = BASE_DIR + "DATASET/OUTPUT/"
elif os_name == 'posix':
    BASE_DIR = "/media/pinkalinux/WORK/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/"
    INPUT_DIR = BASE_DIR + "DATASET/INPUT/"
    OUTPUT_DIR = BASE_DIR + "DATASET/OUTPUT/"

SEED = 6886

%matplotlib inline

## Import data

# Load document list
skill_tokens_long_df = pd.read_csv(OUTPUT_DIR + "skill_tokens_long_df.csv")
skill_tokens_long_df.iloc[:25, ]

skill_tokens_long_df[["skill_token"]] = skill_tokens_long_df[["skill_token"]].applymap(
    ast.literal_eval
)
skill_tokens_long_df

# Load half-size reduced word embeddings
word_embeddings = KeyedVectors.load(OUTPUT_DIR + "skill_halfsize_word_norm_vectors.kv")

## Create document embeddings

def vectorize(key_list, doc_list, kv, vector_size):
    """Generate vectors for list of documents using a Word Embedding

    Args:
        key_list: List of unique ids (job_id + skill_id) of documents
        doc_list: List of documents
        kv: GloVe's Word Embedding in Gensim Keyed Vectors format
        vector_size: dimension of 1 word embedding

    Returns:
        Dictionary of document ids and vectors (mean of word vectors)
    """
    features = {}
    zero_vector = np.zeros(vector_size)

    process_doc_list = tqdm(enumerate(doc_list))
    process_doc_list.set_description(desc="Creating document embeddings")

    for idx, tokens in process_doc_list:
        key_id = key_list[idx]
        vectors = []
        for token in tokens:
            if kv.has_index_for(token):
                vectors.append(kv.get_vector(token))
            else:
                continue
        if vectors:
            vectors = np.asarray(vectors)
            avg_vec = vectors.mean(axis=0)
            features[key_id] = avg_vec
        else:
            features[key_id] = zero_vector
    return features

vector_size = word_embeddings.vector_size
vector_size

# Create document embeddings for skills from job description
skill_description_dict = vectorize(
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_description", "key_id"
    ].values.tolist(),
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_description", "skill_token"
    ].values.tolist(),
    word_embeddings,
    vector_size,
)

# Save document embeddings for skills from job description in file
skill_description_dict_outfile = (
    OUTPUT_DIR + "skill_description_halfsize_long_dict.pkl"
)
# np.save(skill_description_embeddings_outfile, skill_description_embeddings)
pickle.dump(
    skill_description_dict, open(skill_description_dict_outfile, "wb")
)

# Create document embeddings for skills from job requirement
skill_requirement_dict = vectorize(
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_requirement", "key_id"
    ].values.tolist(),
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_requirement", "skill_token"
    ].values.tolist(),
    word_embeddings,
    vector_size,
)

# Save document embeddings for skills from job requirement in file
skill_requirement_dict_outfile = (
    OUTPUT_DIR + "skill_requirement_halfsize_long_dict.pkl"
)
pickle.dump(
    skill_requirement_dict, open(skill_requirement_dict_outfile, "wb")
)

# # Create document embeddings for skills from job qualification
skill_qualification_dict = vectorize(
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_qualification", "key_id"
    ].values.tolist(),
    skill_tokens_long_df.loc[
        skill_tokens_long_df["column_type"] == "job_qualification", "skill_token"
    ].values.tolist(),
    word_embeddings,
    vector_size,
)

# Save document embeddings for skills from job qualification in file
skill_qualification_dict_outfile = (
    OUTPUT_DIR + "skill_qualification_halfsize_long_dict.pkl"
)
pickle.dump(
    skill_qualification_dict, open(skill_qualification_dict_outfile, "wb")
)

# Create empty array to store embeddings of 3 columns
description_keys = set(skill_description_dict.keys())
requirement_keys = set(skill_requirement_dict.keys())
qualification_keys = set(skill_qualification_dict.keys())
key_sets = description_keys | requirement_keys | qualification_keys

doc_length = len(key_sets)
doc_width = vector_size * 3
skill_wide_embeddings = np.zeros((doc_length, doc_width))
doc_length, doc_width

# Concat embeddings of 3 columns with corresponding key_id (job_id + skill_id)
skill_embeddings_dict = {}
for idx, key in tqdm(enumerate(sorted(key_sets))):
    if key in description_keys:
        vector1 = skill_description_dict[key]
        skill_wide_embeddings[idx, :vector_size] = vector1
        row_key = str(key) + '01'
        skill_embeddings_dict[row_key] = vector1
    if key in requirement_keys:
        vector2 = skill_requirement_dict[key]
        skill_wide_embeddings[idx, vector_size:(vector_size*2)] = vector2
        row_key = str(key) + '02'
        skill_embeddings_dict[row_key] = vector2
    if key in qualification_keys:
        vector3 = skill_qualification_dict[key]
        skill_wide_embeddings[idx, (vector_size*2):] = vector3
        row_key = str(key) + '03'
        skill_embeddings_dict[row_key] = vector3

skill_wide_embeddings.shape, len(skill_embeddings_dict)

skill_wide_embeddings[:2, :20]

# Save document embeddings of skill concatenating in numpy format file
skill_wide_embeddings_outfile = OUTPUT_DIR + "skill_wide_halfsize_embeddings.npz"
np.savez_compressed(
    skill_wide_embeddings_outfile, skill_wide_embeddings=skill_wide_embeddings
)

pickle.dump(
    skill_embeddings_dict,
    open(OUTPUT_DIR + "skill_halfsize_embeddings_dict.pkl", "wb"),
)
