==> CODES/10_latex_converter.Rmd <==
---
title: "Untitled"
author: "Thien Binh Hoang Dao"
date: "11/05/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

```{r cars}
summary(cars)
```

$$
g_j = \dfrac{(n-j)ng_{j-1} + \sum_{i=1}^{n-j+1} (|c_{i} - c_{s_{1}} - c_{s_{2}}| - |c_{i} - c_{s_{1}}| - |c_{i} - c_{s_{2}}|) - c_{s_{2}} - c_{s_{1}} + |c_{s_{1}} - c_{s_{2}}|}{(n-j-1)n}
$$

## Including Plots

You can also embed plots, for example:

```{r pressure, echo=FALSE}
library(Hmisc)
library(DT)
library(dplyr)


base_path <- "E:\\THIENDHB_GOOGLEDRIVE\\MASTER TILBURG\\THESIS\\"
df <-
  read.csv(
    file.path(
      base_path,
      'RESULTS',
      "skill_halfsize_hopkins_test_result.csv"
    ),
    stringsAsFactors = F
  )
df$dataset <- 'skill\\_halfsize\\_embeddings'
df <-
  df %>% rename(
    "n\\_samples" = n,
    "hopkins\\_score" = hopkins_stastitics,
    "dataset\\_size" = dataset_size,
    "test\\_duration" = test_duration
  )
df["hopkins\\_score"] <- round(df["hopkins\\_score"], 4)
datatable(df)

latex_df <- Hmisc::latex(
  df,
  caption = 'Hopkins scores computed on skill embeddings',
  label = 'tbl:hopkins_skill', where = "!htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/hopkins_skill', ".tex", sep =
                 ""
  )
)

```

``` {r}

df <-
  read.csv(
    file.path(
      base_path,
      'RESULTS',
      "title_hopkins_test_result.csv"
    ),
    stringsAsFactors = F
  )
df$dataset <- 'title\\_embeddings'
df <-
  df %>% rename(
    "n\\_samples" = n,
    "hopkins\\_score" = hopkins_stastitics,
    "dataset\\_size" = dataset_size,
    "test\\_duration" = test_duration
  )
df["hopkins\\_score"] <- round(df["hopkins\\_score"], 4)
datatable(df)

latex_df <- Hmisc::latex(
  df,
  caption = 'Hopkins scores computed on title embeddings',
  label = 'tbl:hopkins_title', where = "!htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/hopkins_title', ".tex", sep =
                 ""
  )
)

```

``` {r}

library(readxl)
hyper_tuning_space <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "space"
  )

latex_df <- Hmisc::latex(
  hyper_tuning_space,
  caption = 'Hyperparameter search space',
  label = 'tbl:hyper_search_space', where = "!htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/hyper_search_space', ".tex", sep =
                 ""
  )
)

```

``` {r}

optimal_title <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "optimal-title"
  )

latex_df <- Hmisc::latex(
  optimal_title,
  caption = 'Optimal hyperparameter for job title clustering',
  label = 'tbl:optimal_hyper_title', where = "!htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/optimal_hyper_title', ".tex", sep =
                 ""
  )
)

```

``` {r}

optimal_skill <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "optimal-title"
  )

latex_df <- Hmisc::latex(
  optimal_skill,
  caption = 'Optimal hyperparameter for skill set clustering',
  label = 'tbl:optimal_hyper_skill', where = "!htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/optimal_hyper_skill', ".tex", sep =
                 ""
  )
)

```

``` {r}
library(stringr)
model_names <- c('kmeans', 'genie', 'hdbscan', 'lda', 'gsdmm')
emb_names <- c('title', 'skill')

for (m in model_names) {
  for (i in emb_names) {
    try ({
      df <-
        read.csv(file.path(
          base_path,
          'RESULTS',
          m,
          paste0(i, "_", m, "_tuning_result.csv")
        ),
        stringsAsFactors = F)
      
      colnames <- colnames(df)
      new_colnames <- str_replace_all(colnames, '\\_', ' \\\\_')
      colnames(df) <- new_colnames
      removed_colnames <-
        new_colnames[str_detect(new_colnames, 'error')]
      removed_colnames <- c('duration', 'silhouette \\_corr', removed_colnames)
      kept_colnames <-
        new_colnames[!new_colnames %in% removed_colnames]
      round_colnames <-
        c(
          'inertia',
          'silhouette',
          'calinski \\_harabasz',
          'davies \\_bouldin',
          'coherence \\_umass',
          'coherence \\_cv',
          'perplexity'
        )
      df <- df[kept_colnames]
      for (c in round_colnames) {
        if (c %in% colnames(df)) {
          df[c] <- round(df[c], 4)
        }
      }
      
      # rows <- nrow(df)
      # colnames_length <- str_length(final_colnames)
      # names_sum <- sum(colnames_length)
      # col_ratios <- colnames_length / names_sum
      col_length <- sapply(df, function(x) max(nchar(x), na.rm = T), simplify=T)
      char_sum <- sum(col_length)
      col_ratios <- col_length / char_sum
      # col_ratio1 <- paste0(">{\\RaggedLeft\\hspace{0pt}}m{", col_ratios[1], "\\linewidth}")
      
      # cgroup_adjustment <- paste0("{>{\\Centering\\hspace{0pt}}m{", col_ratios, "\\linewidth}}")
      col_adjustment <- paste0(">{\\RaggedLeft\\hspace{0pt}}m{", col_ratios, "\\linewidth}")
      collabel_adjustment <- paste0(">{\\Centering\\hspace{0pt}}m{", col_ratios, "\\linewidth}")
      
      latex_df <- Hmisc::latex(
        df,
        caption = paste0(
          'Hyperparameter tuning results of ',
          m,
          ' for ',
          i ,
          ' clustering'
        ),
        label = paste0('tbl:', m, '_hyper_tuning_', i),
        where = "!htbp",
        rowname = NULL,
        col.just = col_adjustment, 
        collabel.just = collabel_adjustment,
        lines.page = 35, 
        longtable = T,
        booktabs = T,
        # ctable = T,
        multicol = T,
        file = file.path(
          base_path,
          'RESULTS',
          paste0(i, "_", m, "_tuning_result.tex")
        )
      )

      
      # \begin{tabular}{rrrrrr}
      # \end{tabular}\end{center}
      
      # tmp <- readLines(file.path(
      #     base_path,
      #     'RESULTS',
      #     paste0(i, "_", m, "_tuning_result.tex")
      #   ))
      # start_row <- tmp[grepl("\\\\begin\\{tabular\\}", tmp)]
      # tmp[grepl("\\\\begin\\{tabular\\}", tmp)] <- paste("\\begin{adjustbox}{width=\\linewidth}", start_row[1])
      # tmp[grepl("\\\\end\\{tabular\\}\\\\end\\{center\\}", tmp)] <- "\\end{tabular}\\end{adjustbox}\\end{center}"
      # writeLines(tmp, file.path(
      #     base_path,
      #     'RESULTS',
      #     paste0(i, "_", m, "_tuning_result.tex")
      #   ))
      
    })
  }
}

```

``` {r}

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "comparison1-title"
  )

# df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)

latex_df <- Hmisc::latex(
  df,
  caption = 'Performance comparison between title clustering models',
  label = 'tbl:title_performance_comparison', where = "htbp",
  # dec = 4,
  cdec = c(0,0,4,4,4,4,4),
  rowname = NULL,
  file = paste(base_path, '/RESULTS/title_performance_comparison', ".tex", sep =
                 ""
  )
)

```

``` {r}

# df <-
#   read_excel(
#     "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
#     sheet = "comparison2-title"
#   )
# 
# df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)
# 
# latex_df <- Hmisc::latex(
#   df,
#   caption = 'Performance comparison between title topic models',
#   label = 'tbl:title_topic_performance_comparison', where = "htbp",
#   rowname = NULL,
#   file = paste(base_path, '/RESULTS/title_topic_performance_comparison', ".tex", sep =
#                  ""
#   )
# )

```

``` {r}

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "comparison1-skill"
  )

# df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)

latex_df <- Hmisc::latex(
  df,
  caption = 'Performance comparison between skill clustering models',
  label = 'tbl:skill_performance_comparison', where = "htbp",
  cdec = c(0,0,4,4,4,4,4),
  rowname = NULL,
  file = paste(base_path, '/RESULTS/skill_performance_comparison', ".tex", sep =
                 ""
  )
)

```

``` {r}

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/hyper_tuning_space.xlsx",
    sheet = "comparison2-skill"
  )

df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)

latex_df <- Hmisc::latex(
  df,
  caption = 'Performance comparison between skill topic models',
  label = 'tbl:skill_topic_performance_comparison', where = "htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/skill_topic_performance_comparison', ".tex", sep =
                 ""
  )
)

```

``` {r}

library(Hmisc)
library(readxl)

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/lda_topic_demonstration.xlsx",
    sheet = "lda-demo"
  )

# df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)

latex_df <- Hmisc::latex(
  df,
  caption = 'Examples of LDA topic modelling',
  label = 'tbl:lda_demo_example', where = "htbp",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/lda_demo_example', ".tex", sep =
                 ""
  )
)

```


``` {r}

library(Hmisc)
library(readxl)

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/lda_topic_demonstration.xlsx",
    sheet = "lda-topic-keywords"
  )

# df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)

latex_df <- Hmisc::latex(
  df,
  caption = 'Topic keywords of LDA clustering for job titles',
  label = 'tbl:title_lda_topic_keywords', where = "ht!",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/title_lda_topic_keywords', ".tex", sep =
                 ""
  )
)

```

``` {r}

library(Hmisc)
library(readxl)

indices <- c(1, 2, 3)
model <- c('kmeans', 'genie', 'hdbscan',)
model_names <- c('k-means', 'Genie', 'HDBSCAN')

for (i in indices) {
  df <-
    read_excel(
      "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/topic_demonstration.xlsx",
      sheet = paste0("title-", model[i], "-topwords")
    )
  
  # df[c(2:ncol(df))] <- round(df[c(2:ncol(df))], 4)
  
  latex_df <- Hmisc::latex(
    df,
    caption = paste0('Topic keywords of ', model_names[i], ' clustering for job titles'),
    label = paste0('tbl:title_', model[i], '_topwords'), where = "ht!",
    rowname = NULL,
    file = paste(base_path, paste0('/RESULTS/title_', model[i], '_topwords'), ".tex", sep =
                   ""
    )
  )
}

```

``` {r}

library(stringr)
library(readxl)

model_names <- c('kmeans', 'genie', 'hdbscan', 'lda', 'gsdmm')
emb_names <- c('title', 'skill')

for (m in model_names) {
  for (i in emb_names) {
    try ({
      df <-
        read_excel(
          "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/topic_demonstration.xlsx",
          sheet = paste0(i, "-", m, "-topwords")
        )
      
      col_length <- sapply(df, function(x) max(nchar(x), na.rm = T), simplify=T)
      char_sum <- sum(col_length)
      col_ratios <- col_length / char_sum
      # col_ratio1 <- paste0(">{\\RaggedLeft\\hspace{0pt}}m{", col_ratios[1], "\\linewidth}")
      
      # cgroup_adjustment <- paste0("{>{\\Centering\\hspace{0pt}}m{", col_ratios, "\\linewidth}}")
      col_adjustment <- paste0(">{\\RaggedLeft\\hspace{0pt}}m{", col_ratios, "\\linewidth}")
      collabel_adjustment <- paste0(">{\\Centering\\hspace{0pt}}m{", col_ratios, "\\linewidth}")
      
      latex_df <- Hmisc::latex(
        df,
        caption = paste0('Topic keywords of ', m, ' clustering for job ', i, 's'),
        label = paste0('tbl:', i ,'_', m, '_topwords'), where = "ht!",
        col.just = col_adjustment,
        collabel.just = collabel_adjustment,
        multicol= T,
        longtable = T,
        booktabs = T,
        rowname = NULL,
        lines.page = 35, 
        file = paste(base_path, paste0('/RESULTS/', i ,'_', m, '_topwords'), ".tex", sep =
                       ""
        ))
    })
  }
}


```

``` {r}

library(Hmisc)
library(readxl)

df <-
  read_excel(
    "E:/THIENDHB_GOOGLEDRIVE/MASTER TILBURG/THESIS/RESULTS/topic_demonstration.xlsx",
    sheet = "urf_feature_ranking"
  )

latex_df <- Hmisc::latex(
  df,
  caption = '\\gls{urf} feature importance ranking',
  label = 'fig:urf_feature_importance', where = "ht!",
  rowname = NULL,
  file = paste(base_path, '/RESULTS/urf_feature_ranking', ".tex", sep =
                 ""
  )
)

```
